moving some model stuff
