+++ app/src/main/java/com/damsoft/overheidsdata/apimodel/Facets.java
@@ -1,0 +1,7 @@
+package com.damsoft.overheidsdata.apimodel;
+
+
+public class Facets {
+
+
+}
