+++ app/src/main/java/com/damsoft/overheidsdata/apimodel/SearchFacets.java
@@ -1,0 +1,7 @@
+package com.damsoft.overheidsdata.apimodel;
+
+
+public class SearchFacets {
+
+
+}
