--- app/src/main/java/com/damsoft/overheidsdata/apimodel/packages/SearchFacets.java
@@ -1,7 +1,0 @@
-package com.damsoft.overheidsdata.apimodel.packages;
-
-
-public class SearchFacets {
-
-
-}
